fix(catalog): keep RequireMD mounted across call-letter clear cycles
The peek control's validity guard sat above RequireMD, so clearing the code_letters field unmounted the authorization gate and forced a full org-role re-resolution on every clear-and-retype cycle, along with a blank flicker while it re-resolved. Move the guard inside RequireMD's child so authority state survives input transitions, trim code_letters before checking it so whitespace-only input can't slip through, and debounce code_letters and genre_id together as one composed value so a genre change mid-typing can't pair with letters the user hasn't finished composing. Pass preferCacheValue to the lazy trigger so a genre-dropdown flip-flop reuses an already-cached response instead of re-hitting the backend.

Add a live region and an explicit label association to the preview so screen readers announce updates and errors instead of silence, and relocate the component out of ArtistAdd/ into catalog/ directly since it has no coupling to that unbuilt form beyond stale documentation.

Test fixes: the two "renders nothing" tests asserted their negative before RequireMD's authority resolution had actually settled, passing vacuously regardless of the guard; await resolution first. Add coverage for debounce coalescing (proven red against a debounce removal) and for the stale-response race between a slow and a fast query, and draw call-letter and genre literals from the shared test constants instead of inlining them.
